Messed up the terms sample and frame
A frame is a sample on every channel. A sample is the most little unit
for pcm data (two bytes).
TransformSound works correctly now.

+++ test/org/toilelibre/soundtransform/SlowdownTest.java
@@ -10,7 +10,7 @@ public class SlowdownTest {
 
 	@Test
 	public void test (){
-		double [] testarray = new double [] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15};
+		long [] testarray = new long [] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15};
 		Sound testsound = new Sound (testarray, 1, testarray.length);
 		SlowdownSoundTransformation est =
 				new SlowdownSoundTransformation (2);
